make tags page metadata not article specific for CPS assets
