+++ src/app/containers/ArticleMetadata/index.jsx
@@ -1,11 +1,6 @@
 import React from 'react';
-import { string, shape, arrayOf, objectOf, any } from 'prop-types';
+import { string, objectOf, any } from 'prop-types';
 import Metadata from '../Metadata';
-
-const renderTags = tags =>
-  tags.map(({ thingLabel: content }) => (
-    <meta name="article:tag" content={content} key={content} />
-  ));
 
 const ArticleMetadata = ({
   articleId,
@@ -25,26 +20,15 @@ const ArticleMetadata = ({
       lang={lang}
       description={description}
       openGraphType="article"
+      aboutTags={aboutTags}
+      mentionsTags={mentionsTags}
     >
       <meta name="article:author" content={author} />
       <meta name="article:modified_time" content={lastPublished} />
       <meta name="article:published_time" content={firstPublished} />
       {section && <meta name="article:section" content={section} />}
-      {Boolean(aboutTags && aboutTags.length) && renderTags(aboutTags)}
-      {Boolean(mentionsTags && mentionsTags.length) && renderTags(mentionsTags)}
     </Metadata>
   );
-
-const tagPropTypes = shape({
-  thingUri: string,
-  topicId: string,
-  topicName: string,
-  curationType: arrayOf(string),
-  thingId: string,
-  thingLabel: string,
-  thingType: arrayOf(string),
-  thingSameAs: arrayOf(string),
-});
 
 ArticleMetadata.propTypes = {
   title: string.isRequired,
@@ -52,8 +36,6 @@ ArticleMetadata.propTypes = {
   firstPublished: string.isRequired,
   lastPublished: string.isRequired,
   section: string,
-  aboutTags: arrayOf(tagPropTypes),
-  mentionsTags: arrayOf(tagPropTypes),
   lang: string.isRequired,
   description: string.isRequired,
   linkedData: objectOf(any),
@@ -61,8 +43,6 @@ ArticleMetadata.propTypes = {
 
 ArticleMetadata.defaultProps = {
   section: '',
-  aboutTags: [],
-  mentionsTags: [],
   linkedData: {},
 };
