+++ src/app/pages/CpsMap/index.jsx
@@ -48,6 +48,7 @@ const CpsMapContainer = ({ pageData }) => {
   );
   const firstPublished = getFirstPublished(pageData);
   const lastPublished = getLastPublished(pageData);
+  const aboutTags = path(['metadata', 'tags', 'about'], pageData);
 
   const componentsToRender = {
     fauxHeadline,
@@ -75,6 +76,7 @@ const CpsMapContainer = ({ pageData }) => {
         lang={metadata.language}
         description={summary}
         openGraphType="website"
+        aboutTags={aboutTags}
       >
         <meta name="article:published_time" content={firstPublished} />
         <meta name="article:modified_time" content={lastPublished} />
@@ -86,6 +88,7 @@ const CpsMapContainer = ({ pageData }) => {
         showAuthor
         datePublished={firstPublished}
         dateModified={lastPublished}
+        aboutTags={aboutTags}
       />
       <ATIAnalytics data={pageData} />
       <StyledGhostGrid as="main" role="main">
